Removed myAccount and refactored it into profile with a button when logged in in the dropdown menu
