+++ index.html
@@ -29,9 +29,6 @@
             <li class="nav-item">
                 <a href="#admin" class="nav-link">Admin</a>
             </li>
-            <li class="nav-item">
-                <a href="#myAccount" class="nav-link">My Account</a>
-            </li>
             <!-- Account Dropdown -->
             <li class="nav-item dropdown" id="accountDropdownContainer">
                 <!-- Dynamic content will be injected here -->
@@ -41,9 +38,8 @@
 </nav>
 
 <!-- Dynamic Content Area -->
-<div id="app" class="container">
-    <!-- Content will be injected here -->
-</div>
+<div id="app" class="container"></div>
+
 
 <!-- jQuery, Popper.js, and Bootstrap JS -->
 <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
